- when allowRGroups and allowOptionalAttachments are true, bonds connecting R groups to the scaffold should match single or aromatic (#6306)
- added relevant unit tests

+++ Code/MinimalLib/tests/tests.js
@@ -823,7 +823,6 @@ M  END
 
 function test_has_coords() {
     var mol = RDKitModule.get_mol('CC');
-    console.log(`1) test_has_coords`);
     assert(!mol.has_coords());
     var mol2 = RDKitModule.get_mol(mol.get_new_coords());
     assert(mol2.has_coords() === 2);
@@ -854,7 +853,7 @@ function test_has_coords() {
   3  9  1  0
 M  END
 `);
-assert(mol3.has_coords() === 3);
+    assert(mol3.has_coords() === 3);
 }
 
 function test_kekulize() {
@@ -1794,8 +1793,9 @@ M  END
 }
 
 function test_query_colour() {
-    var mol = RDKitModule.get_qmol('c1ccc2nc([*:1])nc([*:2])c2c1');
+    var mol;
     try {
+        mol = RDKitModule.get_qmol('c1ccc2nc([*:1])nc([*:2])c2c1');
         var svg1 = mol.get_svg_with_highlights(JSON.stringify({width: 350, height: 300}));
         assert(svg1.includes("width='350px'"));
         assert(svg1.includes("height='300px'"));
@@ -1807,7 +1807,88 @@ function test_query_colour() {
         assert(svg2.includes("</svg>"));
         assert(!svg2.includes("#7F7F7F"));
     } finally {
-        mol.delete();
+        if (mol) {
+            mol.delete();
+        }
+    }
+}
+
+function test_alignment_r_groups_aromatic_ring() {
+    var mol;
+    var scaffold;
+    try {
+        mol = RDKitModule.get_mol('c1ccc2nccnc2c1');
+        assert(mol && mol.is_valid());
+        scaffold = RDKitModule.get_mol(`
+  MJ201100                      
+
+  8  8  0  0  0  0  0  0  0  0999 V2000
+   -1.0263   -0.3133    0.0000 R#  0  0  0  0  0  0  0  0  0  0  0  0
+   -2.4553    0.5116    0.0000 R#  0  0  0  0  0  0  0  0  0  0  0  0
+   -1.7408   -0.7258    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+   -1.7408   -1.5509    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+   -2.4553   -1.9633    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+   -3.1698   -1.5509    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+   -3.1698   -0.7258    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+   -2.4553   -0.3133    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+  3  1  1  0  0  0  0
+  8  2  1  0  0  0  0
+  4  3  2  0  0  0  0
+  5  4  1  0  0  0  0
+  6  5  2  0  0  0  0
+  7  6  1  0  0  0  0
+  8  3  1  0  0  0  0
+  8  7  2  0  0  0  0
+M  RGP  2   1   2   2   1
+M  END`);
+        assert(scaffold && scaffold.is_valid());
+        var res = mol.generate_aligned_coords(scaffold, JSON.stringify({useCoordGen: true, allowRGroups: true}));
+        assert(res);
+        assert.equal(JSON.parse(res).atoms.length, 8);
+        assert.equal(JSON.parse(res).bonds.length, 8);
+    } finally {
+        if (mol) {
+            mol.delete();
+        }
+    }
+    try {
+        mol = RDKitModule.get_mol(`
+  MJ201100                      
+
+ 10 11  0  0  0  0  0  0  0  0999 V2000
+    3.6937    2.5671    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    2.8687    2.5671    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    2.4561    1.8526    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    2.8687    1.1382    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    3.6937    1.1381    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    4.1062    1.8526    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    4.9313    1.8527    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0
+    5.3438    2.5671    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    4.9313    3.2816    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0
+    4.1062    3.2816    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0
+  5  6  1  0  0  0  0
+  4  5  2  0  0  0  0
+  3  4  1  0  0  0  0
+  2  3  2  0  0  0  0
+  1  6  2  0  0  0  0
+  1  2  1  0  0  0  0
+  8  9  1  0  0  0  0
+  9 10  2  0  0  0  0
+ 10  1  1  0  0  0  0
+  7  8  2  0  0  0  0
+  6  7  1  0  0  0  0
+M  END`);
+        var res = mol.generate_aligned_coords(scaffold, JSON.stringify({allowRGroups: true, alignOnly: true}));
+        assert(res);
+        assert.equal(JSON.parse(res).atoms.length, 8);
+        assert.equal(JSON.parse(res).bonds.length, 8);
+    } finally {
+        if (mol) {
+            mol.delete();
+        }
+        if (scaffold) {
+            scaffold.delete();
+        }
     }
 }
 
@@ -1864,6 +1945,7 @@ initRDKitModule().then(function(instance) {
     test_get_frags();
     test_hs_in_place();
     test_query_colour();
+    test_alignment_r_groups_aromatic_ring();
     waitAllTestsFinished().then(() =>
         console.log("Tests finished successfully")
     );
